:lipstick: refactor handleAuthentication definition (#56)
This commit removes the second parameter of the handleAuthentication
method, as it is never used
and refactor the first argument to use ES6 destructuring for clear
description of
what is the subject of the method implementation. The nextState has been
replaced
with location object directly as the subject of the method check

Thanks!